requirements.txt
torch>=1.10.0
torchvision>=0.10.0
Pillow>=8.0.0
opencv-python>=4.5.0
matplotlib>=3.5.0
numpy>=1.18.0
streamlit
